requirements.txt
torch
transformers
torchvision
pillow
pandas
tqdm
scikit-learn
seaborn
matplotlib
streamlit
seaborn